W3: hoist Kit specs + kitPrompt to core (engine-consumable); ui re-exports, components stay
